02_src/webhook-handler/requirements.txt
# Cloud Functions
functions-framework==3.*

# LINE Messaging API
line-bot-sdk==3.*

# PostgreSQL
psycopg2-binary==2.9.*
sqlalchemy==2.0.*
pg8000==1.*

# Google Cloud
google-cloud-secret-manager==2.*
cloud-sql-python-connector==1.*
google-cloud-tasks==2.*

# AI - New Google Gen AI SDK
google-genai>=0.2.0

# Utilities
python-dotenv==1.*

# Web Scraping
requests==2.*
beautifulsoup4==4.*

# Payment
stripe==10.*
